+++ sand_scripts/predefined_datasets.py
@@ -1,0 +1,5 @@
+from sm_datasets.datasets import Datasets
+
+
+def load(dataset: str):
+    return Datasets().get_dataset(dataset)