+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # Changelog
 
+## [4.1.0] - 2024-04-13
+
+### Added
+
+- SAND CLI can import dataset returned from a python function
+
+### Changed
+
+- SAND CLI create project command does not require project's description (default to use dataset name)
+
 ## [Unreleased](https://github.com/usc-isi-i2/sand/tree/HEAD)
 
 [Full Changelog](https://github.com/usc-isi-i2/sand/compare/2.1.5...HEAD)
 
-**Changed**
+### Changed
 
 - Update dependency: `d-repr` version constraint to `^2.10.0` from prerelease version.
 - Update dockerfile and add docker-compose example.